Add a proof-of-concept of a landing page (#18)
Part of a resolution for
https://github.com/openUC2/imswitch-os/issues/13 .

The landing page added by this PR is provided by
https://github.com/openUC2/device-portal (a new repo created for this
PR), which uses https://github.com/PlanktoScope/device-portal as starter
code. This landing page doesn't actually provide any of the links
specified in https://github.com/openUC2/imswitch-os/issues/13; that will
be done in a future PR. This landing page is only a minimal
proof-of-concept which strips down whatever was irrelevant from
PlanktoScope/device-portal.