Support `openuc2.local` & `uc2.local` as mDNS names (#19)
This PR follows up on
https://github.com/openUC2/imswitch-os/issues/14#issuecomment-3284422065
by adding support for `openuc2.local` and `uc2.local` as simpler
machine-independent aliases for the `openuc2-{machine-name}.local`
domain name. These are mDNS domain names, so they should usually work
from a client web browser which is on the same network router or
Tailscale network as the RPi, even when the client web browser's device
isn't directly connected to the RPi (via the RPi's Wi-Fi hotspot or via
an Ethernet cable directly to the RPi). However, they won't work in web
browsers installed on Linux via Flatpak, because of
https://github.com/flatpak/flatpak/issues/348 &
https://github.com/flatpak/xdg-desktop-portal/discussions/1365 (which
are extremely unlikely to get resolved in the foreseeable future,
unfortunately).

This work is tracked in
https://www.notion.so/Support-machine-generic-mDNS-names-2754e612c78a8058a9dcff2d4dd69d6b?source=copy_link