fix: improve the use of ImSwitch Docker container images (#22)
Previously, the pallet specified an arm64-only Docker container image
for ImSwitch, which meant that an unmodified pallet couldn't be tested
on an x64-64-based computer. This PR builds upon
https://github.com/openUC2/ImSwitch/pull/174 by switching to a
multiplatform container image whose CPU architecture variant is
automatically selected. As a result, I'm now able to test the pallet on
my amd64-based laptop.

This PR also switches from the `latest` tag on that container image to a
stable tag which must be manually changed in order to use a newer
version of the container image (cc @beniroquai because this probably
affects your developer experience), in order to make pallet deployments
reproducible (because they are very much not reproducible with the
`latest` tag for the ImSwitch container image! With the `latest` tag,
deploying the same pallet tomorrow could give different results from
deploying that pallet today, depending on whether someone pushes up new
commits to https://github.com/openUC2/ImSwitch in the meantime!). To
make it easier to work with non-`latest` container image tags, I will
also make it easier to bump the tag to the newest tagged version in PR
branches (see my TODOs below), and I'll also make it easier to use
`latest` in a branch for testing.

TODOs:

- [x] Also add a feature flag named `dev-edge` in the imswitch package
to use the `edge` tag of the container image (which corresponds to the
latest commit on the master branch) instead of a fixed version tag
- [x] Use a less bloated Docker container image, as built by
https://github.com/openUC2/ImSwitch/pull/175

Ideas for future PRs:

- Add a CI check in PRs to ensure that we don't merge PRs with the
`dev-edge` feature flag activated into the `main` branch, or with the
`latest`/`edge` tag in the `compose.yml` file; also add a CI check in
the `main` branch in case anyone commits directly to `main`
- Add an updatecli workflow which can be manually triggered to bump the
ImSwitch container image in particular, maybe as a new commit on a
target branch
- Add an updatecli workflow which runs monthly(?) to bump all container
images in all Compose apps

This work is tracked in
https://www.notion.so/Add-versioned-tags-to-ImSwitch-s-container-images-2724e612c78a8062aff2f70ebfecf961?source=copy_link
and
https://www.notion.so/Unify-ImSwitch-s-Docker-container-images-for-automatic-selection-of-host-platform-2794e612c78a80898da7db756d087e83?source=copy_link