Vendor pkgs & pkg depls from PlanktoScope pallet (#32)
This pallet was initialized as a customization layer over
https://github.com/PlanktoScope/pallet-standard. However, in the time
since this pallet's initialization, Forklift has been removed for the
next release cycle of PlanktoScope OS - so PlanktoScope/pallet-standard
is no longer maintained. This PR removes imports of packages and package
deployments from PlanktoScope/pallet-standard, and instead vendors them
as files within this pallet itself; thus, all maintenance of those files
will be done in this pallet and not in PlanktoScope/pallet-standard.

With this PR, I've also reorganized the packages and package deployments
so that packages live next to their corresponding package deployments
(i.e. in `/deployments` or a subdirectory of `/deployments`) for better
[locality](https://wiki.c2.com/?LocalityOfReferenceDocumentation), and
so that the directory hierarchy of packages is slightly flattened.
Packages are also renamed to have a `.pkg` suffix for easier treatment
with the file import mechanism.

This work is tracked on Notion at
https://www.notion.so/Migrate-files-from-PlanktoScope-pallet-standard-into-openUC2-pallet-2724e612c78a80e497dcc41a46c2c217?source=copy_link

--

This PR also attempts to fix a nondeterministic bug in which the machine
name sometimes appears to be generated (and saved to
`/run/machine-name`) as "unknown"; my hypothesis is that this bug is
related to ImSwitch OS's use of `C.UTF-8` as the system language instead
of `en_US.UTF-8` (also see
https://github.com/PlanktoScope/pallet-standard/pull/86).