Import OS image build system from openUC2/imswitch-os (#39)
This PR merges https://github.com/openuc2/imswitch-os into the pallet,
to remove the decoupling between OS image builds and the pallet.
Instead, when we commit a change to a pallet, GitHub Actions will
automatically build an OS image for that commit without any further
manual actions required. This removes the toil of having to update a
`forklift-pallet-version` file in a separate repository in order to get
a new OS image for a new commit of the pallet.

This work is tracked on Notion at
https://www.notion.so/Merge-imswitch-os-repo-into-openUC2-pallet-2724e612c78a806a83caf300ab47b422?source=copy_link
.