Rename the repo from openUC2/pallet to openUC2/rpi-imswitch-os (#40)
This PR follows up on #39 and my archival of
https://github.com/openUC2/imswitch-os by renaming this repo from
github.com/openUC2/pallet to github.com/openUC2/rpi-imswitch-os, in
order to reflect that this repo is now responsible for ImSwitch OS. The
pallet is one part of the OS, and it is intended to be the majority of
this repo. Another reason for this PR is that we may eventually want to
create other pallets (e.g. a pallet without ImSwitch, as described in
https://www.notion.so/DN-25-Customer-specific-OS-pre-configuration-2924e612c78a80b9b97cfffc199f3ae5?source=copy_link),
which would be other repos with their own names. This is why we're
changing the name of this repo away from github.com/openUC2/pallet.