Refactor OS build scripts and USB drive automounting (#44)
This PR:

- Moves USB drive automounting out of the OS build scripts (which are
hard to update) and into the Forklift pallet (which is easy to update
and iterate on) instead, and also simplifies the USB drive automounting
mechanism by reimplementing that functionality using
[UDisks2](https://github.com/storaged-project/udisks).
- Removes some parts of the OS build scripts which either 1) appear to
be unused or 2) should not be used.
- Integrates changes I had made in
https://github.com/PlanktoScope/PlanktoScope/pull/613
- Adds a package + package deployment (disabled by default) to open a
few firewall ports (namely 3000, 3001, 3002, and 3003) for testing apps
running directly on the OS (e.g. if the app was built on my laptop and I
`scp`'d it over to the RPi to try running it there; this is what I do
for developing `device-admin`)
- Removes three ImSwitch ports (3000, 8002, 22) from the ImSwitch Docker
Compose App definition which are probably no longer needed (cc
@beniroquai)
- Removes an ImSwitch port (8002) from the firewall settings: it's no
longer needed since https://github.com/openUC2/ImSwitch/pull/187 was
merged

This work is tracked in
https://www.notion.so/Remove-unnecessary-modules-in-the-base-OS-build-scripts-2724e612c78a8083a953e0c890e6495c?source=copy_link
and
https://www.notion.so/Support-saving-data-to-an-external-hard-drive-29c4e612c78a80d29953eece8eba5b86?source=copy_link.